tradingbot_ibkr/requirements.txt
ccxt==3.0.72
ib_insync==0.10.14
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.2.4
Flask==2.4.0
river==0.11.0
scikit-learn==1.4.2
joblib==1.3.2
pywebview==3.8
pystray==0.21.4
Pillow==10.1.0
pywin32==306
tqdm==4.67.1
python-binance==1.0.16
google-cloud-storage==2.11.0
requests==2.31.0
